==> analysis/03-DESeq2_analysis.Rmd <==
---
title: "03-DESeq2_analysis"
author: "Valentin Marteau"
params:
  data: "../data"
  lib:  "../lib"
  results:  "../results"
output:
  html_notebook:
    theme: spacelab
    highlight: textmate
    toc: yes
    number_sections: true
    toc_depth: 3
    toc_float: true
---

```{r, results = "hide"}
# Load required packages
library(conflicted)
library(tidyverse)
library(DESeq2)
library(IHW)
library(EnhancedVolcano)

conflict_prefer("select", "dplyr")
conflict_prefer("filter", "dplyr")

# Load custom functions
source(file.path(params$lib, "run_DESeq2.R"))
```

```{r}
# Read data into memory
count_mat <- readRDS(file = file.path(params$results, "produced_data/01-count_mat.rds"))
metadata <- readRDS(file = file.path(params$results, "produced_data/01-metadata.rds"))
key <- readRDS(file = file.path(params$results, "produced_data/01-key.rds"))
kegg_gene_sets <- readRDS(file = file.path(params$results, "produced_data/01-kegg_gene_sets.rds"))
```

```{r}
# Choose model design
model <- ~condition

# List contrasts for DESeq2
contrast <- list(c("condition", "Ctrl", "APTK"),
                 c("condition", "Ctrl", "APTKA"),
                 c("condition", "APTK", "APTKA"))

res_list <- run_DESeq2(count_mat, metadata, model, contrast, key)

saveRDS(res_list, file = file.path(params$results, "produced_data/03-res_list.rds"))
```

# Have a look at some summary stats
```{r}
# Print number of significant genes
lapply(res_list, function(res) {
  c("< 1%" = sum(res$padj < 0.01, na.rm = TRUE),
    "< 5%" = sum(res$padj < 0.05, na.rm = TRUE))
  })
```
```{r}
# Histogram of p values
lapply(res_list, function(res) {
  hist(res$pvalue, breaks = 20, col = "grey50", border = "white")
  })
```

# Make volcano plots and label metabolism related genes
```{r}
make_volcano <- function(de_res, kegg, nr_labels, pCutoff, FCcutoff, prefix) {

# Get subset of metabolic genes in DESeq2 results
  goi <- de_res |> filter(padj < pCutoff , entrez_id %in% kegg) |> top_n(nr_labels) |> pull(symbol)
  
  p <- EnhancedVolcano(
    toptable = de_res,
    lab = de_res$symbol,
    selectLab = goi,
    x = "log2FoldChange",
    y = "padj",
    pCutoff = pCutoff,
    FCcutoff = FCcutoff,
    drawConnectors = TRUE,
    title = paste0(prefix, "_volcano_plot_genes_of_interest"),
    caption = paste0("fold change cutoff: ", FCcutoff, ", adj.p-value cutoff: ", pCutoff))
  
  ggsave(file.path(params$results, paste0("figures/", prefix, "_volcano_padj_GoI.pdf")), plot = p, width = 297, height = 210, units = "mm")
}

lapply(seq_along(res_list), function(i){
  make_volcano(res_list[[i]], kegg = unique(unlist(kegg_gene_sets)),
               nr_labels = 20, pCutoff = 0.05, FCcutoff = 2,
               prefix = names(res_list[i]))
})
```

# Plot individual genes
```{r}
# Set ggplot basic theme
theme_set(
  theme(panel.grid = element_blank(),
        text = element_text(size = 26),
        plot.title = element_text(size = 16, hjust = 0.5),
        axis.line = element_line(),
        panel.background = element_blank(),
        panel.border = element_blank(),
        panel.grid.minor = element_line(linetype = "dotted"))
  )
```
```{r}
dds <- DESeqDataSetFromMatrix(countData = count_mat,
                              colData = metadata,
                              design = ~condition)

# Visualize the counts for a particular gene
goi <- "Slc16a3"
gene <- res_list[[1]] |> filter(symbol == goi) |> pull("gene_id")
data <- plotCounts(dds, gene = gene,
                   intgroup = c("condition"), returnData = TRUE)

ggplot(data, aes(x = condition, y = count)) +
  scale_y_log10() +
  geom_point(aes(fill = condition), size = 8,
             shape = 21, position = position_jitter(width = 0.1, height = 0)) +
  ggtitle(goi)

# Center data points instead of random jitter
ggplot(data, aes(x = condition, y = count, fill = condition)) +
  scale_y_log10() + 
  geom_dotplot(dotsize = 2, binaxis = "y", stackdir = "center") +
  ggtitle(goi)
```
